+++ src/app/pages/auth/auth.component.html
@@ -4,8 +4,9 @@
             class="w-full lg:w-1/2 lg:flex hidden bg-[url('assets/images/AuthBackground.png')] bg-cover bg-center">
         </div>
 
-        <div class="lg:w-1/2 flex w-full justify-between overflow-y-auto">
-            <div class="relative top-0 left-0 m-6 ">
+        <div
+            class="lg:w-1/2 flex w-full justify-center overflow-y-auto relative">
+            <div class="absolute top-0 left-0 m-6 ">
                 <a class="invisible">
                     <div
                         class="hover:bg-gray-100 rounded-md hover:cursor-pointer">
@@ -43,7 +44,7 @@
                 </div>
             </div>
 
-            <div class="relative top-0 right-0 m-6 hover:cursor-pointer"
+            <div class="absolute top-0 right-0 m-6 hover:cursor-pointer"
                 [routerLink]="['/']">
                 <div class="hover:bg-gray-100 rounded-md hover:cursor-pointer">
                     <svg
