+++ src/app/pages/auth/components/forgot-password/forgot-password.component.html
@@ -1,0 +1,37 @@
+<form
+    class="flex flex-col justify-center items-start gap-8 text-[#666] text-[16px] font-normal"
+    [formGroup]="resetPasswordForm"
+    (submit)="submitForm()">
+
+    <p class="text-[32px] font-medium text-[#333]">Passwort zurücksetzen</p>
+    <div class="flex flex-col items-start gap-6 w-[473px]">
+        <div class="flex flex-col items-start gap-1 self-stretch">
+            <label for="email" class="gap-[10px] self-stretch items-center">E-Mail</label>
+            <input
+                class="w-full text-[20px] h-[56px] p-[10px] border rounded-xl focus:outline-none hover:border-gray-400 focus:border-black"
+                id="email"
+                type="email"
+                formControlName="email"
+                required />
+
+        </div>
+    </div>
+    <div>
+        <button
+            class="flex h-[49.5px] px-[29.5px] py-[35px] items-center gap-[13px] rounded-[150px] bg-[#11253E] hover:bg-[#22364F]">
+            <p
+                class="flex flex-col justify-center  h-[45px] text-white text-[22.5px] font-medium text-center tracking-[0.15px]">
+                E-Mail senden
+            </p>
+            <svg width="10" height="15" viewBox="0 0 10 15" fill="none"
+                class="mt-[3px]"
+                xmlns="http://www.w3.org/2000/svg">
+                <path fill-rule="evenodd" clip-rule="evenodd"
+                    d="M4.93934 7.74999L0 2.81065L2.12132 0.689331L9.18198 7.74999L2.12132 14.8107L0 12.6893L4.93934 7.74999Z"
+                    fill="white" />
+            </svg>
+        </button>
+
+    </div>
+
+</form>
